app/VoiceToText/pyproject.toml
[tool.poetry]
name = "acit4040-ai-project"
version = "0.1.0"
description = ""
authors = ["Jackson Herbert Sinamenye <s371140@oslomet.no>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pytest = "^7.4.2"
numpy = "^1.24.0"
numba = "^0.57.1"
tqdm = "^4.66.1"
more-itertools = "^10.1.0"
tiktoken = "0.3.3"
openai-whisper = "^20230918"
ffmpeg = "^1.4"
pytest-mock = "^3.11.1"
triton = [
    { version = "^2.0.0", platform = "linux" },
    { url = "https://huggingface.co/r4ziel/xformers_pre_built/resolve/main/triton-2.0.0-cp310-cp310-win_amd64.whl", platform = "windows" }
]
setuptools-rust = "^1.7.0"
torch="^2.0.1"
typing-extensions = "4.8.0"
urllib3 = "2.0.5"
werkzeug = "2.3.7"
Flask = "^2.1"
sounddevice="^0.4.6"
cffi = "^1.15.1"
setuptools = "68.2.2"
soundfile = "^0.11.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
